Algorithmic Trading/Binance-API-Tutorial/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
dateparser==1.0.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
pycodestyle==2.7.0
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
